Finish CTA button problem
Was having an issue where the button border wasn't showing up. In CSS, I
did a type selector and set the button to have no border for the header,
but then I directly targeted the button for CTA to have a white border, but
it still wasn't showing up. I thought directly targeting was more important
than type selecting (which it is). The issue was that the border property
is shorthand for width, style, and color, and I didn't have the style. Once
I added that everything was fine.